my-ai-chatbot/backend/requirements.txt
# FastAPI and async web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration management
pydantic>=2.0,<3.0
pydantic-settings==2.0.3

# LLM and langchain frameworks
langchain==0.2.14
langchain-google-genai==1.0.10
langchain-community==0.2.12
langgraph>=0.1.0

# Vector database
chromadb>=0.5.0

# Google AI
google-generativeai>=0.3.0

# Environment variables
python-dotenv==1.0.0

# Document processing
pypdf==6.8.0
pillow>=7.1.0,<10.0.0

# HTTP client
requests==2.31.0

# Frontend
streamlit==1.28.1
